% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aws.R
\name{bucket_contents}
\alias{bucket_contents}
\title{List contents of a S3 bucket}
\usage{
bucket_contents(bucket)
}
\arguments{
\item{bucket}{}
}
\value{
vector of data.frame contents
}
\description{
List contents of a S3 bucket
}
\examples{
\dontrun{
my_bucket <- get_bucket("bucketName")
bucket_contents(my_bucket)
}
}

