% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aws.R
\name{load_from_s3}
\alias{load_from_s3}
\title{Load an object from S3}
\usage{
load_from_s3(object_name, bucket)
}
\arguments{
\item{object_name}{name of object to be retrieved}

\item{bucket}{a s3_bucket object or character string of such}
}
\description{
Load an object from S3
}

