% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snowball-package.R
\docType{package}
\name{snowball}
\alias{snowball}
\alias{snowball-package}
\title{snowball: A package for creating and running a remote cluster}
\description{
Starts an instance as a head node, then spawns worker instances for calculations. Created at the 2016 ROpenSci #auunconf.
}
\section{cluster functions}{

The cluster functions ...
}

\section{interrogation functions}{

The interrogation functions ...
}

