% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_create_cluster}
\alias{snowball_create_cluster}
\title{Creates an empty cluster}
\usage{
snowball_create_cluster(JobType, Resources, Description, AddressId,
  KmsKeyARN, RoleARN, SnowballType, ShippingOption, Notification,
  ForwardingAddressId)
}
\arguments{
\item{JobType}{[required] The type of job for this cluster. Currently, the only job type supported
for clusters is \code{LOCAL_USE}.}

\item{Resources}{[required] The resources associated with the cluster job. These resources include
Amazon S3 buckets and optional AWS Lambda functions written in the
Python language.}

\item{Description}{An optional description of this specific cluster, for example
\code{Environmental Data Cluster-01}.}

\item{AddressId}{[required] The ID for the address that you want the cluster shipped to.}

\item{KmsKeyARN}{The \code{KmsKeyARN} value that you want to associate with this cluster.
\code{KmsKeyARN} values are created by using the
\href{http://docs.aws.amazon.com/kms/latest/APIReference/API_CreateKey.html}{CreateKey}
API action in AWS Key Management Service (AWS KMS).}

\item{RoleARN}{[required] The \code{RoleARN} that you want to associate with this cluster. \code{RoleArn}
values are created by using the
\href{http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateRole.html}{CreateRole}
API action in AWS Identity and Access Management (IAM).}

\item{SnowballType}{The type of AWS Snowball device to use for this cluster. The only
supported device types for cluster jobs are \code{EDGE}, \code{EDGE_C}, and
\code{EDGE_CG}.}

\item{ShippingOption}{[required] The shipping speed for each node in this cluster. This speed doesn't
dictate how soon you'll get each Snowball Edge device, rather it
represents how quickly each device moves to its destination while in
transit. Regional shipping speeds are as follows:
\itemize{
\item In Australia, you have access to express shipping. Typically,
devices shipped express are delivered in about a day.
\item In the European Union (EU), you have access to express shipping.
Typically, Snowball Edges shipped express are delivered in about a
day. In addition, most countries in the EU have access to standard
shipping, which typically takes less than a week, one way.
\item In India, devices are delivered in one to seven days.
\item In the US, you have access to one-day shipping and two-day shipping.
}}

\item{Notification}{The Amazon Simple Notification Service (Amazon SNS) notification
settings for this cluster.}

\item{ForwardingAddressId}{The forwarding address ID for a cluster. This field is not supported in
most regions.}
}
\description{
Creates an empty cluster. Each cluster supports five nodes. You use the
CreateJob action separately to create the jobs for each of these nodes.
The cluster does not ship until these five node jobs have been created.
}
\section{Request syntax}{
\preformatted{snowball$create_cluster(
  JobType = "IMPORT"|"EXPORT"|"LOCAL_USE",
  Resources = list(
    S3Resources = list(
      list(
        BucketArn = "string",
        KeyRange = list(
          BeginMarker = "string",
          EndMarker = "string"
        )
      )
    ),
    LambdaResources = list(
      list(
        LambdaArn = "string",
        EventTriggers = list(
          list(
            EventResourceARN = "string"
          )
        )
      )
    ),
    Ec2AmiResources = list(
      list(
        AmiId = "string",
        SnowballAmiId = "string"
      )
    )
  ),
  Description = "string",
  AddressId = "string",
  KmsKeyARN = "string",
  RoleARN = "string",
  SnowballType = "STANDARD"|"EDGE"|"EDGE_C"|"EDGE_CG",
  ShippingOption = "SECOND_DAY"|"NEXT_DAY"|"EXPRESS"|"STANDARD",
  Notification = list(
    SnsTopicARN = "string",
    JobStatesToNotify = list(
      "New"|"PreparingAppliance"|"PreparingShipment"|"InTransitToCustomer"|"WithCustomer"|"InTransitToAWS"|"WithAWSSortingFacility"|"WithAWS"|"InProgress"|"Complete"|"Cancelled"|"Listing"|"Pending"
    ),
    NotifyAll = TRUE|FALSE
  ),
  ForwardingAddressId = "string"
)
}
}

\examples{
# Creates an empty cluster. Each cluster supports five nodes. You use the
# CreateJob action separately to create the jobs for each of these nodes.
# The cluster does not ship until these five node jobs have been created.
\donttest{snowball$create_cluster(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b",
  Description = "MyCluster",
  JobType = "LOCAL_USE",
  KmsKeyARN = "arn:aws:kms:us-east-1:123456789012:key/abcd1234-12ab-34cd-56ef-123456123456",
  Notification = list(
    JobStatesToNotify = list(),
    NotifyAll = FALSE
  ),
  Resources = list(
    S3Resources = list(
      list(
        BucketArn = "arn:aws:s3:::MyBucket",
        KeyRange = structure(
          list(),
          .Names = character(
            0
          )
        )
      )
    )
  ),
  RoleARN = "arn:aws:iam::123456789012:role/snowball-import-S3-role",
  ShippingOption = "SECOND_DAY",
  SnowballType = "EDGE"
)}

}
\keyword{internal}
