% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runCloudAWS.R
\name{execCloudAWS}
\alias{execCloudAWS}
\title{Execute cloud script}
\usage{
execCloudAWS(script, S3_BUCKET = "auunconfdt", S3_SCRIPT = "script.R",
  S3_DATA_OUT = "data.Rdata", S3_LOG_OUT = "out.txt", ...)
}
\arguments{
\item{script}{R script to be uploaded to workers}

\item{S3_BUCKET}{name of bucket}

\item{S3_SCRIPT}{name of script to run on workers}

\item{S3_DATA_OUT}{}

\item{S3_LOG_OUT}{}

\item{...}{}
}
\description{
Execute cloud script
}

