% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aws.R
\name{snowball_setup}
\alias{snowball_setup}
\title{Perform snowball setup}
\usage{
snowball_setup(config_file = "~/.snowball", echo = TRUE)
}
\arguments{
\item{config_file}{location of the configuration file, default ~/.snowball}
}
\value{
NULL if echo=FALSE, otherwise prints config details
}
\description{
Perform snowball setup
}
\examples{
\dontrun{
snowball_setup()
}
}

